requirements.txt
# 桌面应用框架
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0

# Word 文档处理
python-docx>=1.1.0
docxtpl>=1.1.0

# 数据处理
python-dateutil>=2.8.2

# 工具库
pathlib2>=2.3.7; python_version < '3.4'

